feat : Comment 등록기능 구현

+++ src/main/java/com/app/jungsuri/domain/comment/model/Comment.java
@@ -17,7 +17,7 @@ public class Comment {
 
     private final String author;
 
-    private final Long likeCount;
+    private final Integer likeCount;
 
     private final LocalDateTime createdAt;
